Hacked version that relies on respawning
This code is supposed to use the JointStateWrapper class to accept reset
messages that shutdown the JointStatePublisher, reload the
robot_description data from the ROS param server, then restart
publishing. Something isn't right with that, so I hacked the code to
shutdown on the reset message, which (sorta) does the right thing when
its restarted.